Make links work when they are equivalent
There's a bug somewhere else in the code that causes the link in the truncation string to be added to the tail of the string many many times (every drawRect). That bug makes it so at this point it's common for the active link to not be the same instance as the link found to be at that point. This fixes the symptom by allowing the tap to progress forward as long as the underlying URLs are the same.